checking for NULL

+++ 0x13-bit_manipulation/4-clear_bit.c
@@ -11,7 +11,7 @@ int clear_bit(unsigned long int *n, unsigned int index)
 
 	mask = 1;
 	mask = mask << index;
-	if (index > sizeof(unsigned int) * 8 || n == NULL)
+	if (index > sizeof(unsigned long int) * 8 || n == NULL)
 		return (-1);
 	if (((*n >> index) & 1) == 1) /*Checks if 1 at position*/
 		*n = mask ^ *n;
